Move networking keywords to network.py and network.ts
