+++ Browser/keywords/__init__.py
@@ -3,6 +3,7 @@ from .evaluation import Evaluation
 from .playwright_state import PlaywrightState
 from .getters import Getters
 from .input import Input
+from .network import Network
 from .waiter import Waiter
 from .webapp_state import WebAppState
 
@@ -11,6 +12,7 @@ __all__ = [
     "Getters",
     "Evaluation",
     "Input",
+    "Network",
     "PlaywrightState",
     "Waiter",
     "WebAppState",
