+++ node/playwright-wrapper/evaluation.ts
@@ -2,7 +2,7 @@ import { sendUnaryData, ServerUnaryCall } from 'grpc';
 import { Page } from 'playwright';
 import { v4 as uuidv4 } from 'uuid';
 
-import * as pb from './generated/playwright_pb';
+import { Request, Response } from './generated/playwright_pb';
 import { invokeOnPage, invokePlaywirghtMethod, waitUntilElementExists } from './playwirght-invoke';
 import { emptyWithLog, jsResponse, stringResponse } from './response-util';
 import { PlaywrightState } from './playwright-state';
@@ -19,8 +19,8 @@ declare global {
  * RF keywords.
  */
 export async function getElement(
-    call: ServerUnaryCall<pb.Request.ElementSelector>,
-    callback: sendUnaryData<pb.Response.String>,
+    call: ServerUnaryCall<Request.ElementSelector>,
+    callback: sendUnaryData<Response.String>,
     state: PlaywrightState,
 ) {
     await waitUntilElementExists(state, callback, call.request.getSelector());
@@ -31,22 +31,22 @@ export async function getElement(
 }
 
 export async function executeJavascriptOnPage(
-    call: ServerUnaryCall<pb.Request.JavascriptCode>,
-    callback: sendUnaryData<pb.Response.JavascriptExecutionResult>,
+    call: ServerUnaryCall<Request.JavascriptCode>,
+    callback: sendUnaryData<Response.JavascriptExecutionResult>,
     page?: Page,
 ) {
     const result = await invokeOnPage(page, callback, 'evaluate', call.request.getScript());
     callback(null, jsResponse(result));
 }
 
-export async function getPageState(callback: sendUnaryData<pb.Response.JavascriptExecutionResult>, page?: Page) {
+export async function getPageState(callback: sendUnaryData<Response.JavascriptExecutionResult>, page?: Page) {
     const result = await invokeOnPage(page, callback, 'evaluate', () => window.__RFBROWSER__);
     callback(null, jsResponse(result));
 }
 
 export async function waitForElementState(
-    call: ServerUnaryCall<pb.Request.ElementSelectorWithOptions>,
-    callback: sendUnaryData<pb.Response.Empty>,
+    call: ServerUnaryCall<Request.ElementSelectorWithOptions>,
+    callback: sendUnaryData<Response.Empty>,
     state: PlaywrightState,
 ) {
     const selector = call.request.getSelector();
@@ -56,8 +56,8 @@ export async function waitForElementState(
 }
 
 export async function waitForFunction(
-    call: ServerUnaryCall<pb.Request.WaitForFunctionOptions>,
-    callback: sendUnaryData<pb.Response.String>,
+    call: ServerUnaryCall<Request.WaitForFunctionOptions>,
+    callback: sendUnaryData<Response.String>,
     page?: Page,
 ): Promise<void> {
     const script = call.request.getScript();
@@ -69,8 +69,8 @@ export async function waitForFunction(
 }
 
 export async function addStyleTag(
-    call: ServerUnaryCall<pb.Request.StyleTag>,
-    callback: sendUnaryData<pb.Response.Empty>,
+    call: ServerUnaryCall<Request.StyleTag>,
+    callback: sendUnaryData<Response.Empty>,
     page?: Page,
 ) {
     const content = call.request.getContent();
@@ -79,8 +79,8 @@ export async function addStyleTag(
 }
 
 export async function highlightElements(
-    call: ServerUnaryCall<pb.Request.ElementSelectorWithDuration>,
-    callback: sendUnaryData<pb.Response.JavascriptExecutionResult>,
+    call: ServerUnaryCall<Request.ElementSelectorWithDuration>,
+    callback: sendUnaryData<Response.JavascriptExecutionResult>,
     state: PlaywrightState,
 ) {
     const selector = call.request.getSelector();
@@ -103,64 +103,4 @@ export async function highlightElements(
         });
     };
     await invokePlaywirghtMethod(state, callback, '$$eval', selector, highlighter, duration);
-}
-
-export async function httpRequest(
-    call: ServerUnaryCall<pb.Request.HttpRequest>,
-    callback: sendUnaryData<pb.Response.String>,
-    page?: Page,
-) {
-    const opts: { [k: string]: any } = {
-        method: call.request.getMethod(),
-        url: call.request.getUrl(),
-        headers: JSON.parse(call.request.getHeaders()),
-    };
-    if (opts.method != 'GET') {
-        opts.body = call.request.getBody();
-    }
-    try {
-        const response = await page?.evaluate(({ url, method, body, headers }) => {
-            return fetch(url, { method, body, headers }).then((data: Response) => {
-                return data.text().then((body) => {
-                    const headers: { [k: string]: any } = {};
-                    data.headers.forEach((value, name) => (headers[name] = value));
-                    return {
-                        status: data.status,
-                        body: body,
-                        headers: JSON.stringify(headers),
-                        type: data.type,
-                        statusText: data.statusText,
-                        url: data.url,
-                        ok: data.ok,
-                        redirected: data.redirected,
-                    };
-                });
-            });
-        }, opts);
-        callback(null, stringResponse(JSON.stringify(response)));
-    } catch (e) {
-        callback(e, null);
-    }
-}
-
-export async function waitForResponse(
-    call: ServerUnaryCall<pb.Request.HttpCapture>,
-    callback: sendUnaryData<pb.Response.String>,
-    page?: Page,
-) {
-    const urlOrPredicate = call.request.getUrlorpredicate();
-    const timeout = call.request.getTimeout();
-    const result = await invokeOnPage(page, callback, 'waitForResponse', urlOrPredicate, { timeout: timeout });
-    const body = await result.body();
-    callback(null, stringResponse(body));
-}
-export async function waitForRequest(
-    call: ServerUnaryCall<pb.Request.HttpCapture>,
-    callback: sendUnaryData<pb.Response.String>,
-    page?: Page,
-) {
-    const urlOrPredicate = call.request.getUrlorpredicate();
-    const timeout = call.request.getTimeout();
-    const result = await invokeOnPage(page, callback, 'waitForRequest', urlOrPredicate, { timeout: timeout });
-    callback(null, stringResponse(result.url()));
 }
