+++ node/playwright-wrapper/grpc-service.ts
@@ -6,6 +6,7 @@ import * as browserControl from './browser-control';
 import * as evaluation from './evaluation';
 import * as getters from './getters';
 import * as interaction from './interaction';
+import * as network from './network';
 import * as playwrightState from './playwright-state';
 import { PlaywrightState } from './playwright-state';
 
@@ -232,10 +233,10 @@ export class PlaywrightServer implements IPlaywrightServer {
         evaluation.waitForElementState(call, callback, this.state);
     }
     async waitForRequest(call: ServerUnaryCall<Request.HttpCapture>, callback: sendUnaryData<Response.String>) {
-        evaluation.waitForRequest(call, callback, this.getActivePage());
+        network.waitForRequest(call, callback, this.getActivePage());
     }
     async waitForResponse(call: ServerUnaryCall<Request.HttpCapture>, callback: sendUnaryData<Response.String>) {
-        evaluation.waitForResponse(call, callback, this.getActivePage());
+        network.waitForResponse(call, callback, this.getActivePage());
     }
 
     async waitForFunction(
@@ -283,6 +284,6 @@ export class PlaywrightServer implements IPlaywrightServer {
         call: ServerUnaryCall<Request.HttpRequest>,
         callback: sendUnaryData<Response.String>,
     ): Promise<void> {
-        evaluation.httpRequest(call, callback, this.getActivePage());
+        network.httpRequest(call, callback, this.getActivePage());
     }
 }
